backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
mediapipe==0.10.7